Fix problem with double clicking main menu buttons
Prior to this commit double clicking a button on the main menu could
start an activity twice while also breaking functionality of the
activity.
Add credit to @vi for suggesting a mode with a more aggressive computer
Knight button now starts a new game if knights only mode is not active
and vice versa for the solo play button, which now guarantees a normal game.